Added docker file and updated README
