+++ workflow/__init__.py
@@ -1,4 +1,4 @@
 name = "workflow"
 from .multiDimProcess import create_metadata_object, parse_tiff_input, parse_bg_options, \
-    process_background, loopPos, loopT, loopZSm, loopZBg
-from utils.imgProcessing import compute_flat_field
+    process_background, loopPos, loopT, loopZSm, loopZBg, compute_flat_field
+
